Обновление календулы

+++ test/test.cpp
@@ -1,9 +1,12 @@
-/* Автоматизация тестирования MLC компилятора. */
+/* Автоматизация тестирования компилятора Flower. */
 
 #include <iostream>
 #include <fstream>
 #include <cstdlib>
 #include <cstdio>
+#include <vector>
+#include <chrono>
+#include <ctime>
 
 #include <unistd.h>
 #include <signal.h>
@@ -11,6 +14,21 @@
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <fcntl.h>
+
+#include "version.hpp"
+
+std::string gettime(void) {
+    using namespace std::chrono;
+    auto timepoint = system_clock::now();
+    auto h_m_s = system_clock::to_time_t(timepoint);
+    auto ms = time_point_cast<std::chrono::milliseconds>(timepoint);
+
+    char buffer[sizeof "23:59:59.999"];
+    std::snprintf(buffer + std::strftime(buffer, sizeof buffer - 3,
+        "%T.", std::localtime(&h_m_s)), 4, "%03lu", 
+        ms.time_since_epoch().count() % 1000);
+    return buffer;
+}
 
 int compare(std::ifstream & file1, std::ifstream & file2) {
     int ret = 0;
@@ -90,8 +108,8 @@ int genCaseIn(std::ifstream & cases) {
 int runValgrind(std::string filename, bool optimize, std::string input = "") {
     int ret = 0;
 
-    ret += checkMem("./mlc -c " + std::string(optimize ? "-O" : "") + " -i " + 
-                    filename + " -s -o test.bin > /dev/null");
+    ret += checkMem("./flc " + std::string(optimize ? "-O " : "") + 
+                    filename + " test.bin > /dev/null");
     std::cout << filename << std::endl;
 
     if (input != "") {
@@ -102,7 +120,7 @@ int runValgrind(std::string filename, bool optimize, std::string input = "") {
         while (line == "case") {
             if (genCaseIn(cases) != 0)
                 break;
-            ret += checkMem("./mlc -r -s -i test.bin > /dev/null < case.in");
+            ret += checkMem("./flvm test.bin > /dev/null < case.in");
             std::system("rm ./case.in");
             std::cout << "Case #" << caseIterator + 1 << std::endl;
             cases >> line;
@@ -117,9 +135,10 @@ int runValgrind(std::string filename, bool optimize, std::string input = "") {
 
 int runA(std::string filename, std::string input, std::string output, bool optimize) {
     int caseIterator = 0, errorIterator = 0;
+    bool compileError = false;
     std::cout << "[A-unit ";
-    std::system( ("./mlc -c " + std::string(optimize ? "-O" : "") + " -i " +
-                  filename + " -s -o test.bin > a.out").data() );
+    std::system( ("./flc " + std::string(optimize ? "-O " : "") +
+                  filename + " test.bin > a.out").data() );
     std::ifstream log("a.out");
     std::ifstream expected(output);
 
@@ -128,8 +147,10 @@ int runA(std::string filename, std::string input, std::string output, bool optim
     log >> line;
     log.close();
     //#ifndef DEBUG
-    if (line != "КОМПИЛЯЦИЯ:") errorIterator = 1;
-    else {
+    if (line != "КОМПИЛЯЦИЯ:") {
+        errorIterator = 1;
+        compileError = true;
+    } else {
     //#endif
         std::ifstream cases(input);
         cases >> line;
@@ -137,9 +158,9 @@ int runA(std::string filename, std::string input, std::string output, bool optim
             if (genCaseIn(cases) != 0)
                 break;
             //#ifdef DEBUG
-            //std::system("gdb --args ./mlc -r -s -i test.bin");
+            //std::system("gdb --args ./flc -s -i test.bin");
             //#else
-            std::system("./mlc -r -s -i test.bin > a.out < case.in");
+            std::system("./flvm test.bin > a.out < case.in");
             //#endif
             std::system("rm ./case.in");
             std::ifstream actual("a.out");
@@ -168,11 +189,17 @@ int runA(std::string filename, std::string input, std::string output, bool optim
 
     expected.close();
 
-    std::system("rm ./a.out ./test.bin");
+    if (!compileError)
+        std::system("rm ./a.out ./test.bin");
     
     if (caseIterator == 0) {
-        errorIterator = -1;
-        std::cout << " [Тесты не найдены]";
+        if (compileError) {
+            std::cout << " [ Ошибка компиляции ]";
+            errorIterator = -1;
+        } else {
+            std::cout << " [ Тесты не найдены ]";
+            errorIterator = -2;
+        }
     } else errorIterator = errorIterator != 0 ? 1 : 0;
 
     std::cout << std::endl;
@@ -184,7 +211,7 @@ int runA(std::string filename, std::string input, std::string output, bool optim
 int runB(std::string filename, std::string output, bool optimize) {
     int ret = 0;
     std::cout << "[B-unit ";
-    std::system( ("./mlc -c " + std::string(optimize ? "-O" : "") + " -i " +
+    std::system( ("./flc " + std::string(optimize ? "-O " : "") +
                   filename + " > a.out").data() );
     std::ifstream log("a.out");
     std::ifstream expected(output);
@@ -206,12 +233,23 @@ int runB(std::string filename, std::string output, bool optimize) {
     return ret;
 }
 
+void rebuild(const char * pattern, bool make = true) {
+    std::system("cp ./Makefile ./temp");
+    std::system( ("sed -e 's/" + std::string(pattern) + "/g' ./temp > ./Makefile").data() );
+    std::system("rm ./temp");
+    if (make)
+        std::system("make -s clean; make -s");
+}
 
 int main(int argc, char ** argv) {
     int errors = 0, notFound = 0;
     bool mem = false;
     bool opt = false;
     bool bisect = false;
+    bool restore = false;
+    bool coverage = false;
+
+    std::cout << "Автоматизация тестирования Flower " << VERSION << std::endl;
 
     for (int i = argc; i > 1; i--) {
         if (argv[i - 1][0] == '-') {
@@ -225,23 +263,45 @@ int main(int argc, char ** argv) {
                 case 'f':
                     bisect = true;
                     break;
+                case 'r':
+                    restore = true;
+                    break;
+                case 'c':
+                    coverage = true;
+                    break;
                 default:
                     throw std::runtime_error("Неизвестный флаг.");
             }
         }
     }
 
+    std::vector<std::string> filesWithError, filesWithCompileError, testsNotFoundError;
+
     #ifdef DEBUG
     std::cout << "Собрано с отладочной информацией, пересборка.\n";
-    std::system("cp ./Makefile ./temp");
-    std::system("sed -e 's/RELEASE=NO/RELEASE=YES/g' ./temp > ./Makefile");
-    std::system("rm ./temp");
-    std::system("make clean; make");
+    rebuild("RELEASE=NO/RELEASE=YES");
     #endif
 
-    if (bisect) std::system("rm ./bin/*; make mlc");
+    #ifndef TCOV
+    if (coverage) {
+        std::cout << "Пересборка для анализа покрытия.\n";
+        rebuild("COVERAGE=NO/COVERAGE=YES");
+    }
+    #else
+    if (!coverage) {
+        std::cout << "Пересборка без анализа покрытия.\n";
+        rebuild("COVERAGE=YES/COVERAGE=NO");
+    }
+    #endif
 
-    int flags = (int)mem + (int)opt + (int)bisect;
+    if (bisect) std::system("rm ./bin/*; make flc flvm");
+
+    auto start = std::chrono::system_clock::now();
+    
+    std::cout << "Запуск тестирования" << (mem? " с Valgrind" : "");
+    std::cout << " в " << gettime() << "\n\n";
+
+    int flags = (int)mem + (int)opt + (int)bisect + (int)restore + (int)coverage;
     for (int i = flags + 1; i < argc; i++) {
         std::string filename = argv[i];
         bool unitA = filename.find("/A-unit/") != std::string::npos;
@@ -261,23 +321,72 @@ int main(int argc, char ** argv) {
         source = source.substr(0, source.find("."));
         std::string output = path + "output" + source + ".out";
 
+        int x;
+
         if (unitA) {
             std::string input = path + "input" + source + ".in";
-            if (mem) errors += runValgrind(filename, opt, input);
-            else {
-                int x = runA(filename, input, output, opt);
-                if (x > 0) errors += x;
-                else notFound -= x;
-            }
-        } else if (mem) errors += runValgrind(filename, opt); 
-        else errors += runB(filename, output, opt);
+            if (mem) x = runValgrind(filename, opt, input);
+            else x = runA(filename, input, output, opt);
+        } else 
+            if (mem) x = runValgrind(filename, opt); 
+            else x = runB(filename, output, opt);
+
+        if (x > 0) {
+            errors += x;
+            filesWithError.push_back(filename);
+        } else if (x < 0) {
+            notFound += 1;
+            if (x == -1)
+                filesWithCompileError.push_back(filename);
+            else
+                testsNotFoundError.push_back(filename);
+        }
 
         if ((errors != 0) && bisect) break;
     }
 
-    std::cout << "\nПройдено " << argc - ((int)mem + (int)opt + 1) << " тестов, из них:\n\t";
-    std::cout << argc - ((int)mem + (int)opt + 1) - errors - notFound << " успешно\n\t" << errors;
+    
+    if (restore) {
+        #if defined(DEBUG) || defined(TCOV)
+        std::cout << "Возврат параметров Makefile...\n";
+        #endif
+
+        #if defined(DEBUG) && !defined(TCOV)
+        rebuild("RELEASE=YES/RELEASE=NO", false);
+        #endif
+
+        #if defined(TCOV)
+        rebuild("COVERAGE=YES/COVERAGE=NO");
+        #endif
+    }
+
+    int all = argc - (flags + 1);
+    std::cout << "\nПройдено " << all << " тестов, из них:\n\t";
+    std::cout << all - errors - notFound << " успешно\n\t" << errors;
     std::cout << " с ошибкой\n\t" << notFound << " тестов не было запущено\n";
+
+    if (filesWithError.size() > 0)
+        std::cout << "\nОШИБКА НА ТЕСТАХ:\n";
+
+    for (std::string ferr : filesWithError)
+        std::cout << "\t" << ferr << "\n";
+
+    if (filesWithCompileError.size() > 0)
+        std::cout << "\nОШИБКА КОМПИЛЯЦИИ НА ТЕСТАХ:\n";
+
+    for (std::string ferr : filesWithCompileError)
+        std::cout << "\t" << ferr << "\n";
+
+    if (testsNotFoundError.size() > 0)
+        std::cout << "\nТЕСТЫ НЕ НАЙДЕНЫ:\n";
+
+    for (std::string ferr : testsNotFoundError)
+        std::cout << "\t" << ferr << "\n";
+
+    auto end = std::chrono::system_clock::now();
+    auto diff = std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count();
+    std::cout << "\nТест завершён в " << gettime() << "\n";
+    std::cout << "Прошло " << (float) diff / 1000 << " секунд.\n";
 
     return errors + notFound;
 }
